requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
spacy>=3.6.0
sentencepiece>=0.2.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Data processing and storage
pandas>=2.0.0
numpy>=1.24.0
sqlite3  # Built-in with Python
pyarrow>=10.0.0  # For Parquet support

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0
hashlib  # Built-in with Python

# Optional: for better audio support
ffmpeg-python>=0.2.0

# Optional: for advanced data processing
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
